Sarah/first modules (#32)
For issue #24
Implemented Projectile Count module (applied to weapon, increases the amount of projectiles spawned per fire)
Implemented Rate of Fire module; uses T's work, puts Rate of Fire as a variable in WeaponEntity, modifies RateOfFire
Created a few systems to apply the new ship movement speed system as well; forward, backward, side and all.

+++ Assets/Scripts/Scriptables/Data/WeaponData.cs
@@ -25,7 +25,7 @@ namespace Celeritas.Scriptables
 		/// <summary>
 		/// The speed this weapon fires at.
 		/// </summary>
-		public float RateOfFire { get => rateOfFire; }
+		public uint RateOfFire { get => rateOfFire; }
 
 		/// <summary>
 		/// Get if this weapon aims or not.
